requirements.txt
absl-py==1.3.0
altair==4.2.0
asgiref==3.6.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.4
Django==4.1.7
entrypoints==0.4
executing==0.10.0
filelock==3.7.1
Flask==2.1.3
fonttools==4.34.4
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.15.1
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
jupyter-client==7.3.4
jupyter-core==4.11.1
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.23.1
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
protobuf==3.19.6
psutil==5.9.1
psycopg2-binary==2.9.5
pure-eval==0.2.2
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0b4
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.9.3
semver==2.13.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sqlparse==0.4.3
stack-data==0.4.0
streamlit==1.14.0
streamlit-option-menu==0.3.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch-tb-profiler==0.4.0
tornado==6.2
traitlets==5.3.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
virtualenv==20.15.1
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.10.0
